+++ @doc/functions.md
