+++ config/executor/query.go
