+++ query/node_func.go
