+++ query/util.go
